feat: basic working theme on css variables
