--- packages/cookie-theme-toggler/index.ts
@@ -1,1 +1,0 @@
-export * from './cookie-theme-toggler'
