+++ packages/index.ts
@@ -34,6 +34,5 @@ export * from './toast'
 export * from './tooltip'
 export * from './transition'
 export * from './utils'
-export * from './cookie-theme-toggler'
-export * from './content-theme'
+export * from './theme/content-theme'
 export * from './cookies-tooltip'
