+++ packages/theme/content-theme/content-theme.module.css
@@ -1,0 +1,9 @@
+@layer contentTheme {
+  .dark {
+    display: var(--lido-dark-display, contents);
+  }
+
+  .light {
+    display: var(--lido-light-display, contents);
+  }
+}
