+++ packages/theme/content-theme/content-theme.stories.tsx
@@ -1,5 +1,5 @@
 import { StoryFn, Meta } from '@storybook/react'
-import { Block } from '../block'
+import { Block } from '../../block'
 import { ContentTheme } from '.'
 
 export default {
