+++ packages/theme/index.ts
@@ -1,9 +1,11 @@
-export * from './cookie-theme-provider'
+export * from './content-theme'
 export * from './theme-provider'
-export * from './document-head-contents'
-export * from './constants'
+export * from './theme-toggler'
+export * from './cookie-theme-provider'
+export * from './utils/document-head-contents'
 export * from './themes'
-export * from './types'
+export * from './use-theme-toggle'
 export { migrationThemeCookiesToCrossDomainCookiesClientSide } from './utils/cookies-migration'
 
+import './theme-legacy.css'
 import './theme.css'
