+++ packages/theme/theme-provider/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-provider'
