+++ packages/theme/theme-toggler/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-toggler'
