+++ postcss.config.js
@@ -2,7 +2,7 @@
 module.exports = {
   plugins: {
     '@csstools/postcss-global-data': {
-      files: ['packages/theme/base/breakpoints.css'],
+      files: ['packages/theme/breakpoints.css'],
     },
     'postcss-custom-media': {},
     'postcss-nesting': {},
