+++ scripts/verify-allowed-css-variables.mjs
@@ -6,14 +6,20 @@ const getDefinedVariables = async () => {
 
 const getUsedVariables = async () => {
   const files = await glob('packages/**/*.module.css')
-  const usedVars = (await Promise.all(files.map(async file => {
-    const content = await fs.readFile(file, 'utf-8')
-    const matches = content.match(/var\(--lido-[^,)]+\)/g)
-    const vars = matches?.map(rawVariable => rawVariable.match(/--lido-[^,)]+/)?.[0])
-    return vars
-  })))
+  const usedVars = (
+    await Promise.all(
+      files.map(async (file) => {
+        const content = await fs.readFile(file, 'utf-8')
+        const matches = content.match(/var\(--lido-[^,)]+[),]/g)
+        const vars = matches?.map(
+          (rawVariable) => rawVariable.match(/--lido-[^,)]+/)?.[0],
+        )
+        return vars
+      }),
+    )
+  )
     .flat()
-    .filter(result => result != null)
+    .filter((result) => result != null)
   return new Set(usedVars)
 }
 
@@ -21,12 +27,13 @@ const getUsedVariables = async () => {
 const main = async () => {
   const usedVars = await getUsedVariables()
   const definedVars = await getDefinedVariables()
-  // console.log(definedVars)
-  const usedButNotDefined = [...usedVars].filter(usedVar => !definedVars.has(usedVar))
+  const usedButNotDefined = [...usedVars].filter(
+    (usedVar) => !definedVars.has(usedVar),
+  )
   if (usedButNotDefined.length === 0) {
     console.log('All used variables are defined')
     process.exit(0)
-  } 
+  }
 
   for (let weirdVar of usedButNotDefined) {
     console.error(`var(${weirdVar}) used but not defined`)
